+++ typescript/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./models";
 export { logger, configure } from "./logging";
+export { REQUEST_ID_HEADER, resolveCorrelationId, runWithCorrelationId } from "./logging/_correlation";
+export { withRequestLogging } from "./logging/request";